posledni funkcni reseni

+++ Cukrovi/ISurovina.cs
@@ -5,7 +5,7 @@ namespace Cukrovi;
 public interface ISurovina
 {
     string Jmeno { get; }
-    decimal Mnozstvi { get; set; }
+    double Mnozstvi { get; set; }
     string Jednotky { get; set; }
     string Druh { get; set; }
